[refactor] 게시판 클래스들 패키지 리팩터링 #71
게시판 클래스들 패키지 리팩터링
