--- src/main/java/com/sparta/daengtionary/domain/MapComment.java
@@ -1,5 +1,0 @@
-package com.sparta.daengtionary.domain;
-
-
-public class MapComment {
-}
