+++ src/main/java/com/sparta/daengtionary/domain/community/Community.java
@@ -1,25 +1,24 @@
-package com.sparta.daengtionary.domain;
+package com.sparta.daengtionary.domain.community;
 
 import com.sparta.daengtionary.configration.error.CustomException;
 import com.sparta.daengtionary.configration.error.ErrorCode;
-import com.sparta.daengtionary.dto.request.TradeRequestDto;
+import com.sparta.daengtionary.domain.Member;
+import com.sparta.daengtionary.dto.request.CommunityRequestDto;
 import com.sparta.daengtionary.util.Timestamped;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.procedure.spi.ParameterRegistrationImplementor;
 
 import javax.persistence.*;
-import java.time.LocalDateTime;
 
-@Getter
 @Entity
+@Getter
 @NoArgsConstructor
-public class Trade extends Timestamped {
+public class Community extends Timestamped {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long tradeNo;
+    private Long communityNo;
 
     @JoinColumn(name = "mumber_id", nullable = false)
     @ManyToOne(fetch = FetchType.LAZY)
@@ -31,37 +30,22 @@ public class Trade extends Timestamped {
     @Column(nullable = false)
     private String content;
 
-    @Column(nullable = false)
-    private int price;
-
-    @Column(nullable = false)
-    private String status;
-
     @Column
     private int view;
 
     @Builder
-    public Trade(Long tradeNo, Member member, String title, String content, int price){
-        this.tradeNo = tradeNo;
+    public Community(Long communityNo, Member member, String title, String content) {
+        this.communityNo = communityNo;
         this.member = member;
         this.title = title;
         this.content = content;
-        this.price = price;
-        this.status = "판매중";
         this.view = 0;
     }
 
-    public void updateTrade(TradeRequestDto requestDto){
+    public void updateCommunity(CommunityRequestDto requestDto){
         this.title = requestDto.getTitle();
         this.content = requestDto.getContent();
-        this.price = requestDto.getPrice();
-        this.status = requestDto.getStatus();
     }
-
-    public void updateStatus(String status){
-        this.status = status;
-    }
-
     public void viewUpdate() {
         this.view += 1;
     }
@@ -71,6 +55,5 @@ public class Trade extends Timestamped {
             throw new CustomException(ErrorCode.MAP_WRONG_ACCESS);
         }
     }
-
 
 }
