+++ src/main/java/com/sparta/daengtionary/domain/map/Map.java
@@ -1,7 +1,8 @@
-package com.sparta.daengtionary.domain;
+package com.sparta.daengtionary.domain.map;
 
 import com.sparta.daengtionary.configration.error.CustomException;
 import com.sparta.daengtionary.configration.error.ErrorCode;
+import com.sparta.daengtionary.domain.Member;
 import com.sparta.daengtionary.dto.request.MapPutRequestDto;
 import com.sparta.daengtionary.util.Timestamped;
 import lombok.AllArgsConstructor;
