+++ src/main/java/com/sparta/daengtionary/domain/map/MapComment.java
@@ -1,0 +1,5 @@
+package com.sparta.daengtionary.domain.map;
+
+
+public class MapComment {
+}
