+++ src/main/java/com/sparta/daengtionary/domain/trade/Trade.java
@@ -1,23 +1,26 @@
-package com.sparta.daengtionary.domain;
+package com.sparta.daengtionary.domain.trade;
 
 import com.sparta.daengtionary.configration.error.CustomException;
 import com.sparta.daengtionary.configration.error.ErrorCode;
-import com.sparta.daengtionary.dto.request.CommunityRequestDto;
+import com.sparta.daengtionary.domain.Member;
+import com.sparta.daengtionary.dto.request.TradeRequestDto;
 import com.sparta.daengtionary.util.Timestamped;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.hibernate.procedure.spi.ParameterRegistrationImplementor;
 
 import javax.persistence.*;
+import java.time.LocalDateTime;
 
-@Entity
 @Getter
+@Entity
 @NoArgsConstructor
-public class Community extends Timestamped {
+public class Trade extends Timestamped {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long communityNo;
+    private Long tradeNo;
 
     @JoinColumn(name = "mumber_id", nullable = false)
     @ManyToOne(fetch = FetchType.LAZY)
@@ -29,22 +32,37 @@ public class Community extends Timestamped {
     @Column(nullable = false)
     private String content;
 
+    @Column(nullable = false)
+    private int price;
+
+    @Column(nullable = false)
+    private String status;
+
     @Column
     private int view;
 
     @Builder
-    public Community(Long communityNo, Member member, String title, String content) {
-        this.communityNo = communityNo;
+    public Trade(Long tradeNo, Member member, String title, String content, int price){
+        this.tradeNo = tradeNo;
         this.member = member;
         this.title = title;
         this.content = content;
+        this.price = price;
+        this.status = "판매중";
         this.view = 0;
     }
 
-    public void updateCommunity(CommunityRequestDto requestDto){
+    public void updateTrade(TradeRequestDto requestDto){
         this.title = requestDto.getTitle();
         this.content = requestDto.getContent();
+        this.price = requestDto.getPrice();
+        this.status = requestDto.getStatus();
     }
+
+    public void updateStatus(String status){
+        this.status = status;
+    }
+
     public void viewUpdate() {
         this.view += 1;
     }
@@ -54,5 +72,6 @@ public class Community extends Timestamped {
             throw new CustomException(ErrorCode.MAP_WRONG_ACCESS);
         }
     }
+
 
 }
