+++ src/main/java/com/sparta/daengtionary/domain/trade/TradeImg.java
@@ -1,4 +1,4 @@
-package com.sparta.daengtionary.domain;
+package com.sparta.daengtionary.domain.trade;
 
 import lombok.Builder;
 import lombok.Getter;
