+++ src/main/java/com/sparta/daengtionary/dto/response/community/CommunityDetatilResponseDto.java
@@ -1,4 +1,4 @@
-package com.sparta.daengtionary.dto.response;
+package com.sparta.daengtionary.dto.response.community;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Builder;
@@ -6,31 +6,32 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 @Getter
 @NoArgsConstructor
-public class CommunityResponseDto {
-
+public class CommunityDetatilResponseDto {
     private Long communityNo;
     private String nick;
     private String title;
+    private String content;
     private int view;
-    private String communityImg;
+    private List<String> imgList;
     @JsonFormat(pattern = "yy-MM-dd hh:mm:ss")
     private LocalDateTime createdAt;
-
     @JsonFormat(pattern = "yy-MM-dd hh:mm:ss")
-    private  LocalDateTime modifiedAt;
+    private LocalDateTime modifiedAt;
+
     @Builder
-    public CommunityResponseDto(Long communityNo,String nick,String title, int view, String communityImg,
-                                LocalDateTime createdAt, LocalDateTime modifiedAt){
+    public CommunityDetatilResponseDto(Long communityNo, String title, String content, int view,
+                                       String nick , List<String> imgList, LocalDateTime createdAt, LocalDateTime modifiedAt) {
         this.communityNo = communityNo;
         this.nick = nick;
         this.title = title;
+        this.content = content;
         this.view = view;
-        this.communityImg = communityImg;
+        this.imgList = imgList;
         this.createdAt = createdAt;
         this.modifiedAt = modifiedAt;
-
     }
 }
