+++ src/main/java/com/sparta/daengtionary/dto/response/community/CommunityResponseDto.java
@@ -1,4 +1,4 @@
-package com.sparta.daengtionary.dto.response;
+package com.sparta.daengtionary.dto.response.community;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Builder;
@@ -6,32 +6,31 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Getter
 @NoArgsConstructor
-public class CommunityDetatilResponseDto {
+public class CommunityResponseDto {
+
     private Long communityNo;
     private String nick;
     private String title;
-    private String content;
     private int view;
-    private List<String> imgList;
+    private String communityImg;
     @JsonFormat(pattern = "yy-MM-dd hh:mm:ss")
     private LocalDateTime createdAt;
-    @JsonFormat(pattern = "yy-MM-dd hh:mm:ss")
-    private LocalDateTime modifiedAt;
 
+    @JsonFormat(pattern = "yy-MM-dd hh:mm:ss")
+    private  LocalDateTime modifiedAt;
     @Builder
-    public CommunityDetatilResponseDto(Long communityNo, String title, String content, int view,
-                                       String nick , List<String> imgList, LocalDateTime createdAt, LocalDateTime modifiedAt) {
+    public CommunityResponseDto(Long communityNo,String nick,String title, int view, String communityImg,
+                                LocalDateTime createdAt, LocalDateTime modifiedAt){
         this.communityNo = communityNo;
         this.nick = nick;
         this.title = title;
-        this.content = content;
         this.view = view;
-        this.imgList = imgList;
+        this.communityImg = communityImg;
         this.createdAt = createdAt;
         this.modifiedAt = modifiedAt;
+
     }
 }
