+++ src/main/java/com/sparta/daengtionary/dto/response/trade/TradeResponseDto.java
@@ -1,4 +1,4 @@
-package com.sparta.daengtionary.dto.response;
+package com.sparta.daengtionary.dto.response.trade;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Builder;
