+++ src/main/java/com/sparta/daengtionary/repository/community/CommunityRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.daengtionary.repository;
+package com.sparta.daengtionary.repository.community;
 
-import com.sparta.daengtionary.domain.Community;
+import com.sparta.daengtionary.domain.community.Community;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommunityRepository extends JpaRepository<Community,Long> {
