+++ src/main/java/com/sparta/daengtionary/repository/map/MapRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.daengtionary.repository;
+package com.sparta.daengtionary.repository.map;
 
-import com.sparta.daengtionary.domain.Map;
+import com.sparta.daengtionary.domain.map.Map;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MapRepository extends JpaRepository<Map,Long> {
