+++ src/main/java/com/sparta/daengtionary/repository/supportRepository/MapRepositorySupport.java
@@ -4,11 +4,10 @@ package com.sparta.daengtionary.repository.supportRepository;
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.sparta.daengtionary.domain.Map;
-import com.sparta.daengtionary.dto.response.CommunityResponseDto;
-import com.sparta.daengtionary.dto.response.MapResponseDto;
-import com.sparta.daengtionary.dto.response.TradeDetailResponseDto;
-import com.sparta.daengtionary.dto.response.TradeResponseDto;
+import com.sparta.daengtionary.domain.map.Map;
+import com.sparta.daengtionary.dto.response.community.CommunityResponseDto;
+import com.sparta.daengtionary.dto.response.map.MapResponseDto;
+import com.sparta.daengtionary.dto.response.trade.TradeResponseDto;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.support.QuerydslRepositorySupport;
