+++ src/main/java/com/sparta/daengtionary/repository/trade/TradeRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.daengtionary.repository;
+package com.sparta.daengtionary.repository.trade;
 
-import com.sparta.daengtionary.domain.Trade;
+import com.sparta.daengtionary.domain.trade.Trade;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TradeRepository extends JpaRepository<Trade,Long> {
