+++ src/main/java/com/sparta/daengtionary/service/CommunityService.java
@@ -2,17 +2,16 @@ package com.sparta.daengtionary.service;
 
 import com.sparta.daengtionary.configration.error.CustomException;
 import com.sparta.daengtionary.configration.error.ErrorCode;
-import com.sparta.daengtionary.domain.Community;
-import com.sparta.daengtionary.domain.CommunityImg;
+import com.sparta.daengtionary.domain.community.Community;
+import com.sparta.daengtionary.domain.community.CommunityImg;
 import com.sparta.daengtionary.domain.Member;
-import com.sparta.daengtionary.domain.Trade;
 import com.sparta.daengtionary.dto.request.CommunityRequestDto;
-import com.sparta.daengtionary.dto.response.CommunityDetatilResponseDto;
-import com.sparta.daengtionary.dto.response.CommunityResponseDto;
+import com.sparta.daengtionary.dto.response.community.CommunityDetatilResponseDto;
+import com.sparta.daengtionary.dto.response.community.CommunityResponseDto;
 import com.sparta.daengtionary.dto.response.ResponseBodyDto;
 import com.sparta.daengtionary.jwt.TokenProvider;
-import com.sparta.daengtionary.repository.CommunityImgRepository;
-import com.sparta.daengtionary.repository.CommunityRepository;
+import com.sparta.daengtionary.repository.community.CommunityImgRepository;
+import com.sparta.daengtionary.repository.community.CommunityRepository;
 import com.sparta.daengtionary.repository.supportRepository.MapRepositorySupport;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageImpl;
