+++ src/main/java/com/sparta/daengtionary/service/MapService.java
@@ -3,16 +3,18 @@ package com.sparta.daengtionary.service;
 import com.sparta.daengtionary.configration.error.CustomException;
 import com.sparta.daengtionary.configration.error.ErrorCode;
 import com.sparta.daengtionary.domain.*;
+import com.sparta.daengtionary.domain.map.Map;
+import com.sparta.daengtionary.domain.map.MapImg;
+import com.sparta.daengtionary.domain.map.MapInfo;
 import com.sparta.daengtionary.dto.request.MapPutRequestDto;
 import com.sparta.daengtionary.dto.request.MapRequestDto;
-import com.sparta.daengtionary.dto.response.MapDetailResponseDto;
-import com.sparta.daengtionary.dto.response.MapResponseDto;
-import com.sparta.daengtionary.dto.response.MemberResponseDto;
+import com.sparta.daengtionary.dto.response.map.MapDetailResponseDto;
+import com.sparta.daengtionary.dto.response.map.MapResponseDto;
 import com.sparta.daengtionary.dto.response.ResponseBodyDto;
 import com.sparta.daengtionary.jwt.TokenProvider;
-import com.sparta.daengtionary.repository.MapImgRepository;
-import com.sparta.daengtionary.repository.MapInfoRepository;
-import com.sparta.daengtionary.repository.MapRepository;
+import com.sparta.daengtionary.repository.map.MapImgRepository;
+import com.sparta.daengtionary.repository.map.MapInfoRepository;
+import com.sparta.daengtionary.repository.map.MapRepository;
 import com.sparta.daengtionary.repository.supportRepository.MapRepositorySupport;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageImpl;
