+++ src/main/java/com/sparta/daengtionary/service/TradeService.java
@@ -3,14 +3,16 @@ package com.sparta.daengtionary.service;
 import com.sparta.daengtionary.configration.error.CustomException;
 import com.sparta.daengtionary.configration.error.ErrorCode;
 import com.sparta.daengtionary.domain.*;
+import com.sparta.daengtionary.domain.trade.Trade;
+import com.sparta.daengtionary.domain.trade.TradeImg;
 import com.sparta.daengtionary.dto.request.TradeRequestDto;
 import com.sparta.daengtionary.dto.response.ResponseBodyDto;
-import com.sparta.daengtionary.dto.response.TradeDetailResponseDto;
-import com.sparta.daengtionary.dto.response.TradeResponseDto;
+import com.sparta.daengtionary.dto.response.trade.TradeDetailResponseDto;
+import com.sparta.daengtionary.dto.response.trade.TradeResponseDto;
 import com.sparta.daengtionary.jwt.TokenProvider;
 import com.sparta.daengtionary.repository.MemberRepository;
-import com.sparta.daengtionary.repository.TradeImgRepository;
-import com.sparta.daengtionary.repository.TradeRepository;
+import com.sparta.daengtionary.repository.trade.TradeImgRepository;
+import com.sparta.daengtionary.repository.trade.TradeRepository;
 import com.sparta.daengtionary.repository.supportRepository.MapRepositorySupport;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageImpl;
